Terminando la refactorización

+++ models/Adjuntos.php
@@ -19,6 +19,8 @@ class Adjuntos extends \yii\db\ActiveRecord
 {
     const ESCENARIO_FILE = 'file';
 
+    const ESCENARIO_URL = 'url';
+
     public $archivo;
     /**
      * {@inheritdoc}
@@ -37,6 +39,7 @@ class Adjuntos extends \yii\db\ActiveRecord
             [
                 ['url_direccion', 'tarjeta_id',],
                 'required',
+                'on'=>self::ESCENARIO_URL,
             ],
             [['tarjeta_id'], 'default', 'value' => null],
             [['tarjeta_id'], 'integer'],
@@ -149,6 +152,24 @@ class Adjuntos extends \yii\db\ActiveRecord
     }
 
     /**
+     * Devuelve al equipo donde pertenece.
+     * @return Model El modelo de Equipo
+     */
+    public function getEquipo()
+    {
+        return $this->tablero->equipo;
+    }
+
+    /**
+     * Devuelve al tablero donde pertenece.
+     * @return Model Modelo de Tablero.
+     */
+    public function getTablero()
+    {
+        return $this->tarjeta->lista->tablero;
+    }
+
+    /**
      * @return \yii\db\ActiveQuery
      */
     public function getTarjeta()
@@ -167,26 +188,38 @@ class Adjuntos extends \yii\db\ActiveRecord
 
     /**
      * Cuando se elimina el adjunto, se elimina el archivo de
-     * Dropbox.
+     * Dropbox. Se elimina antes de Dropbox y después se elimina
+     * el adjunto.
      * @return [type] [description]
      */
-    public function afterDelete()
+    public function beforeDelete()
     {
-        $cliente = Yii::$app->params['dropbox'];
+        if (!parent::beforeDelete()) {
+            return false;
+        }
 
         if (!$this->esEnlace) {
-
             $extension = substr(
                 $this->nombre,
                 strripos($this->nombre, '.')
             );
-        
+
             UploadFiles::deleteDropbox(
-                $nombre = 'adjunto' . $this->nombre . $this->tarjeta->id
+                $nombre = 'adjunto' . $this->id . $this->tarjeta->id
                     . $extension
             );
         }
 
+        return true;
+    }
+
+    /**
+     * Se crea una notificación de tablero después de eliminar
+     * un adjunto.
+     * @return [type] [description]
+     */
+    public function afterDelete()
+    {
         $tablero = $this->tarjeta->lista->tablero;
         $equipo = $tablero->equipo;
 
@@ -204,26 +237,39 @@ class Adjuntos extends \yii\db\ActiveRecord
         ]))->save();
     }
 
+    /**
+     * Después de guardar el adjunto, si es un archivo, se sube
+     * a Dropbox.
+     * @param  boolean $insert            True si es insert, false si es update.
+     * @param  array   $changedAttributes Atributos antes de modificarse.
+     */
     public function afterSave($insert, $changedAttributes)
     {
-        if (!$insert) {
-            return false;
-        }
-
-        $tablero = $this->tarjeta->lista->tablero;
-        $equipo = $tablero->equipo;
         $tarjeta = $this->tarjeta;
 
-        $miembro = $equipo->getMiembros()
+        $contenido = "ha añadido el adjunto " .
+            (($this->nombre === null) ? $this->url_direccion : $this->nombre) .
+            " en la tarjeta <strong>$tarjeta->denominacion</strong>";
+
+        if (!$insert) {
+            return false;
+            // $contenido = "ha modificado el adjunto " .
+            // (($this->nombre === null) ? $this->url_direccion : $this->nombre) .
+            //     " en la tarjeta $tarjeta->denominacion";
+        }
+
+        $miembro = $this->equipo->getMiembros()
             ->where([
-                'usuario_id'=>Yii::$app->user->id
+                'usuario_id'=>Yii::$app->user->id,
+                'equipo_id'=>$this->equipo->id,
             ])->one();
 
         (new Notificaciones([
-            'contenido'=>"ha añadido un nuevo adjunto" .
-                " en la tarjeta <strong>$tarjeta->denominacion</strong>",
+            'contenido'=>$contenido,
             'miembro_id'=>$miembro->id,
-            'tablero_id'=>$tablero->id,
+            'tablero_id'=>$this->tablero->id,
         ]))->save();
     }
+
+
 }
